+++ src/test/java/io/github/astrapi69/crypt/data/factory/CryptModelFactoryTest.java
@@ -41,7 +41,6 @@ import io.github.astrapi69.crypt.data.key.reader.PrivateKeyReader;
 import io.github.astrapi69.crypt.data.key.reader.PublicKeyReader;
 import io.github.astrapi69.crypt.data.model.CryptModel;
 import io.github.astrapi69.file.search.PathFinder;
-import io.github.astrapi69.meanbean.extension.MeanBeanExtensions;
 
 /**
  * The unit test class for the class {@link CryptModelFactory}
@@ -114,7 +113,8 @@ public class CryptModelFactoryTest
 	@Test
 	public void testWithBeanTester()
 	{
-		MeanBeanExtensions.testWithBeanTester(CryptModelFactory.class);
+		final BeanTester beanTester = new BeanTester();
+		beanTester.testBean(CryptModelFactory.class);
 	}
 
 }