+++ src/test/java/io/github/astrapi69/crypt/data/factory/SecretKeyFactoryExtensionsTest.java
@@ -1,3 +1,27 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 package io.github.astrapi69.crypt.data.factory;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -8,6 +32,7 @@ import javax.crypto.SecretKeyFactory;
 import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanTester;
 
+import io.github.astrapi69.collection.array.ArrayFactory;
 import io.github.astrapi69.crypt.api.algorithm.SunJCEAlgorithm;
 import io.github.astrapi69.crypt.api.algorithm.compound.CompoundAlgorithm;
 
@@ -63,8 +88,10 @@ public class SecretKeyFactoryExtensionsTest
 	@Test
 	public void testNewSecretKeyWithByteArray() throws Exception
 	{
-		byte[] sharedSecret = "sharedSecret".getBytes();
-		String algorithm = "AES";
+		byte[] sharedSecret = ArrayFactory.newByteArray(100, 13, -105, 60, -120, -31, -120, 48, 40,
+			-13, -42, -102, -6, -103, 55, -50, 95, 104, -63, 98, -35, -104, -78, 122, -27, 45, -52,
+			-6, 29, -51, 44, -70);
+		String algorithm = "DES";
 		SecretKey secretKey = SecretKeyFactoryExtensions.newSecretKey(sharedSecret, algorithm);
 		assertNotNull(secretKey);
 	}